+++ src/components/commons/hooks/useCalender.ts
@@ -1,4 +1,4 @@
-import { getDaysInMonth, subMonths } from 'date-fns';
+import { getDaysInMonth } from 'date-fns';
 import { useState } from 'react';
 
 const DATE_MONTH_FIXER = 1;
@@ -48,6 +48,8 @@ const useCalendar = () => {
     },
     []
   );
+
+  console.log(weekCalendarList);
 
   return {
     weekCalendarList: weekCalendarList,
