+++ src/components/commons/layout/footer/layoutFooter.css
@@ -31,7 +31,7 @@ body {
   left: 50%;
   transform: translateX(-50%);
   bottom: 3%;
-  background-color: #391d93;
+  background-color: var(--sub, #8066d1);
 }
 
 .list-item {
