+++ src/components/commons/modals/modalSetting/overlay/myPageOverlay/MyPageOverlay.types.ts
@@ -7,10 +7,6 @@ export interface ValidationMessageProps {
   isError: boolean;
 }
 
-export interface CheckEdit {
-  isEdit: boolean;
-}
-
 export interface IMyPost {
   selectedImage?: File | string;
   username?: string;
