+++ src/components/commons/modals/modalSetting/overlay/myPageOverlay/mypageEdit/MyPageEdit.types.ts
@@ -1,0 +1,13 @@
+export interface CheckEdit {
+  isEdit: boolean;
+}
+
+export interface IMyPost {
+  selectedImage?: File | string;
+  username?: string;
+}
+
+export interface IMyPageEdit {
+  data: any;
+  setIsEdit: React.Dispatch<React.SetStateAction<boolean>>;
+}
